requirements.txt
kafka-python==1.4.1
pyspark==2.3.0
